+++ contracts/mirror_collateral_oracle/src/lib.rs
@@ -5,6 +5,3 @@ pub mod state;
 
 #[cfg(test)]
 mod testing;
-
-#[cfg(test)]
-mod mock_querier;
