+++ contracts/mirror_collateral_oracle/src/testing/mock_querier.rs
