+++ contracts/mirror_collateral_oracle/src/testing/tests.rs
@@ -1,7 +1,7 @@
 use crate::contract::{
     execute, instantiate, query_collateral_info, query_collateral_price, query_config,
 };
-use crate::mock_querier::mock_dependencies;
+use crate::testing::mock_querier::mock_dependencies;
 use cosmwasm_std::testing::{mock_env, mock_info};
 use cosmwasm_std::{Decimal, StdError, Uint128};
 use mirror_protocol::collateral_oracle::{
