+++ contracts/mirror_collector/src/lib.rs
@@ -5,6 +5,3 @@ mod swap;
 
 #[cfg(test)]
 mod testing;
-
-#[cfg(test)]
-mod mock_querier;
