+++ contracts/mirror_collector/src/testing/mod.rs
@@ -1,0 +1,2 @@
+mod mock_querier;
+mod tests;
