+++ contracts/mirror_community/src/lib.rs
@@ -2,4 +2,4 @@ pub mod contract;
 pub mod state;
 
 #[cfg(test)]
-mod testing;
+mod tests;
