+++ contracts/mirror_community/src/tests.rs
@@ -1,5 +1,4 @@
 use crate::contract::{execute, instantiate, query};
-
 use cosmwasm_std::testing::{mock_dependencies, mock_env, mock_info};
 use cosmwasm_std::{from_binary, to_binary, CosmosMsg, StdError, SubMsg, Uint128, WasmMsg};
 use cw20::Cw20ExecuteMsg;
