+++ contracts/mirror_factory/src/lib.rs
@@ -2,11 +2,7 @@ pub mod contract;
 pub mod math;
 pub mod querier;
 pub mod state;
-
 mod response;
 
 #[cfg(test)]
 mod testing;
-
-#[cfg(test)]
-mod mock_querier;
