+++ contracts/mirror_gov/src/lib.rs
@@ -4,7 +4,4 @@ mod staking;
 pub mod state;
 
 #[cfg(test)]
-mod tests;
-
-#[cfg(test)]
-mod mock_querier;
+mod testing;
