+++ contracts/mirror_lock/src/lib.rs
@@ -2,6 +2,4 @@ pub mod contract;
 mod state;
 
 #[cfg(test)]
-mod mock_querier;
-#[cfg(test)]
-mod tests;
+mod testing;
