+++ contracts/mirror_mint/src/lib.rs
@@ -1,16 +1,10 @@
 pub mod contract;
-
 mod asserts;
 mod math;
 mod positions;
 mod querier;
 mod state;
 
-// Testing files
-mod contract_test;
-mod positions_test;
-mod pre_ipo_test;
-mod short_test;
 
 #[cfg(test)]
-mod mock_querier;
+mod testing;
