+++ contracts/mirror_mint/src/testing/mod.rs
@@ -1,0 +1,5 @@
+mod mock_querier;
+mod contract_test;
+mod positions_test;
+mod pre_ipo_test;
+mod short_test;
