+++ contracts/mirror_staking/src/lib.rs
@@ -5,10 +5,5 @@ mod rewards;
 mod staking;
 mod state;
 
-// Testing files
-mod contract_test;
-mod reward_test;
-mod staking_test;
-
 #[cfg(test)]
-mod mock_querier;
+mod testing;
